refactor(asset): extract out loader registry
